Fixed VDS SG list length.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@39559 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/PC/BIOS-new/vds.h
@@ -87,7 +87,7 @@ typedef struct {
 
 /* Higher level routines for utilizing VDS. */
 
-int vds_build_sg_list( vds_edds __far *edds, void __far *buf, uint16_t len );
+int vds_build_sg_list( vds_edds __far *edds, void __far *buf, uint32_t len );
 int vds_free_sg_list( vds_edds __far *edds );
 
 /* Helper for translating 16:16 real mode addresses to 32-bit linear. */
